requirements.txt
pillow==11.2.1
pyHanko==0.26.0
reportlab==4.4.0